Refactors Legendre and SquareRootField
