+++ fields/src/errors/mod.rs
@@ -14,8 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod constraint_field;
+mod constraint_field;
 pub use constraint_field::*;
 
-pub mod field;
+mod field;
 pub use field::*;
