+++ fields/src/legendre.rs
@@ -14,23 +14,25 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-mod field;
-pub use field::*;
+use std::fmt::Debug;
 
-mod fp_parameters;
-pub use fp_parameters::*;
+#[derive(Debug, PartialEq)]
+pub enum LegendreSymbol {
+    Zero = 0,
+    QuadraticResidue = 1,
+    QuadraticNonResidue = -1,
+}
 
-mod one;
-pub use one::*;
+impl LegendreSymbol {
+    pub fn is_zero(&self) -> bool {
+        *self == LegendreSymbol::Zero
+    }
 
-mod prime_field;
-pub use prime_field::*;
+    pub fn is_qnr(&self) -> bool {
+        *self == LegendreSymbol::QuadraticNonResidue
+    }
 
-mod square_root_field;
-pub use square_root_field::*;
-
-mod to_constraint_field;
-pub use to_constraint_field::*;
-
-mod zero;
-pub use zero::*;
+    pub fn is_qr(&self) -> bool {
+        *self == LegendreSymbol::QuadraticResidue
+    }
+}
