+++ fields/src/lib.rs
@@ -54,6 +54,9 @@ pub mod fp6_3over2;
 mod fp12_2over3over2;
 pub use fp12_2over3over2::*;
 
+mod legendre;
+pub use legendre::*;
+
 pub mod tests_field;
 
 mod to_field_vec;
@@ -73,7 +76,6 @@ use snarkvm_utilities::{
         ConstantSerializedSize,
     },
 };
-use std::fmt::Debug;
 
 #[macro_export]
 macro_rules! field {
@@ -98,40 +100,6 @@ macro_rules! field {
             _parameters: std::marker::PhantomData,
         }
     };
-}
-
-/// The interface for a field that supports an efficient square-root operation.
-pub trait SquareRootField: Field {
-    /// Returns the Legendre symbol.
-    fn legendre(&self) -> LegendreSymbol;
-
-    /// Returns the square root of self, if it exists.
-    #[must_use]
-    fn sqrt(&self) -> Option<Self>;
-
-    /// Sets `self` to be the square root of `self`, if it exists.
-    fn sqrt_in_place(&mut self) -> Option<&mut Self>;
-}
-
-#[derive(Debug, PartialEq)]
-pub enum LegendreSymbol {
-    Zero = 0,
-    QuadraticResidue = 1,
-    QuadraticNonResidue = -1,
-}
-
-impl LegendreSymbol {
-    pub fn is_zero(&self) -> bool {
-        *self == LegendreSymbol::Zero
-    }
-
-    pub fn is_qnr(&self) -> bool {
-        *self == LegendreSymbol::QuadraticNonResidue
-    }
-
-    pub fn is_qr(&self) -> bool {
-        *self == LegendreSymbol::QuadraticResidue
-    }
 }
 
 impl_field_into_bigint!(Fp256, BigInteger256, Fp256Parameters);
